adding controllers & services[connecting to DB]

+++ src/main/java/com/example/tpbase/repository/ComputerRepository.java
@@ -3,6 +3,6 @@ package com.example.tpbase.repository;
 import com.example.tpbase.model.Computer;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ComputerRepository extends CrudRepository<Computer, Integer> {
+public interface ComputerRepository extends CrudRepository<Computer, Long> {
 
 }
