+++ src/main/java/com/example/tpbase/repository/DepartementRepository.java
@@ -3,5 +3,5 @@ package com.example.tpbase.repository;
 import com.example.tpbase.model.Departement;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DepartementRepository extends CrudRepository<Departement, Integer> {
+public interface DepartementRepository extends CrudRepository<Departement, Long> {
 }
